improved character movement, started physics
* improved key state management
* removed some redundencies in entity class
* entity movement controlled by physics engine

+++ Goliath/Vector2.h
@@ -10,8 +10,8 @@
 
 struct Vector2
 {
-    int x; //can be positive
-    int y; //can be negative
+    float x; //can be positive
+    float y; //can be negative
 
     Vector2();
     Vector2(float, float);
